SNSアプリ機能完成。CSSの調整及び提出用部分作成予定

+++ LaravelFile/app/Http/Controllers/PostController.php
@@ -8,10 +8,15 @@ use App\Models\Post;
 class PostController extends Controller
 {
 
-    public function indexPost()
+    public function indexPost(Request $request)
     {
+        $userId = $request->query('user_id');
+
         $posts = Post::with('user')
                     ->withCount(['likes', 'comments'])
+                    ->withExists(['likes as is_liked' =>function($query) use ($userId) {
+                        $query->where('user_id', $userId);
+                    }])
                     ->latest()
                     ->get();
         return response()->json($posts);
@@ -27,9 +32,12 @@ class PostController extends Controller
         return response()->json($post, 201);
     }
 
-    public function destroyPost(string $id)
+    public function destroyPost(Request $request, string $id)
     {
         $post = Post::findOrFail($id);
+        if ($post->user_id != $request->query('user_id')) {
+            return response()->json(['message' => '権限がありません'], 403);
+        }
         $post->delete();
         return response()->json(['message' => 'Post deleted successfully']);
     }
